+++ drivers/ethernet/CMakeLists.txt
@@ -29,7 +29,6 @@ zephyr_library_sources_ifdef(CONFIG_ETH_ESP32		eth_esp32.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_LITEX_LITEETH	eth_litex_liteeth.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_SMSC911X	eth_smsc911x.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_STELLARIS	eth_stellaris.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_STM32_HAL	eth_stm32_hal_common.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_W5500		eth_w5500.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_SAM_GMAC	eth_sam_gmac.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_CYCLONEV	eth_cyclonev.c)
@@ -50,6 +49,11 @@ if(CONFIG_ETH_NXP_S32_NETC)
   zephyr_library_sources(eth_nxp_s32_netc.c)
   zephyr_library_sources_ifdef(CONFIG_DT_HAS_NXP_S32_NETC_PSI_ENABLED	eth_nxp_s32_netc_psi.c)
   zephyr_library_sources_ifdef(CONFIG_DT_HAS_NXP_S32_NETC_VSI_ENABLED	eth_nxp_s32_netc_vsi.c)
+endif()
+
+if(CONFIG_ETH_STM32_HAL)
+  zephyr_library_sources(eth_stm32_hal_common.c)
+  zephyr_library_sources_ifdef(CONFIG_PTP_CLOCK_STM32_HAL	eth_stm32_hal_ptp.c)
 endif()
 
 zephyr_library_sources_ifdef(CONFIG_ETH_NXP_S32_GMAC	eth_nxp_s32_gmac.c)
